containersMenu: Remove old items when refreshing menu
Right now items are removed when the menu is opened, before
rendering the new items. Instead, delay the removal until we
are ready to add the new items.

The difference doesn't currently matter, but will when we
fetch containers asynchronously: It is less distracting to
briefly display the previous menu than to flash an empty
menu while the fetch operation is still ongoing.