Re-render when the menu is open
Changes to the containers are now reflected when the menu is open.
Specifically when starting/stopping a container, immediatly when there's
a change the whole menu is redered.

The way its implemented is with 'podman events' command and listening to
the changes in the process inputsteam of that events. Whenever there's
a change on that stream its a sign to rereder the menu.

All the invocations are non-blocking in the UI and the all the events
are processes in the background, never blocking interaction.

See podman events command:
    podman events --filter type=container --format '{\"name\": \"{{ .Name }}\"}'

Signed-off-by: Roy Golan <rgolan@redhat.com>