Support configuring the terminal program
The extension preferences has a new entry for the terminal program to
use for running the commands in a terminal.

The default value is "gnome-terminal --" which is the current behaviour.

Here are terminal values that were tested and worked:
"gnome-terminal --"
"ptyxis --"
"kitty"
"flatpak run com.gexperts.Tilix -e"
"flatpak run app.studiodev.Ptyxis --"

Signed-off-by: Roy Golan <royg21@gmail.com>